+++ src/aui-form-builder-II/js/aui-form-builder-II-field-type.js
@@ -9,6 +9,7 @@
  * A base class for `A.FormBuilderIIFieldType`.
  *
  * @class A.FormBuilderIIFieldType
+ * @extends A.Base
  * @param {Object} config Object literal specifying widget configuration
  *     properties.
  * @constructor