+++ src/aui-form-builder-II/js/aui-form-builder-II.js
@@ -7,18 +7,15 @@
 /**
  * A base class for `A.FormBuilderII`.
  *
- * @class A.FormBuilder
+ * @class A.FormBuilderII
+ * @extends A.Widget
  * @param {Object} config Object literal specifying widget configuration
  *     properties.
  * @constructor
- * @include
- * @include
  */
-A.FormBuilderII  = A.Base.create('form-builder-II', A.Widget, [
-], {
-
+A.FormBuilderII  = A.Base.create('form-builder-II', A.Widget, [], {
     /**
-     * Construction logic executed during the FormBuilder
+     * Construction logic executed during the `A.FormBuilderII`
      * instantiation. Lifecycle.
      *
      * @method initializer
@@ -38,7 +35,8 @@ A.FormBuilderII  = A.Base.create('form-builder-II', A.Widget, [
     },
 
     /**
-     * Destructor lifecycle implementation for the `TooltipDelegate` class.
+     * Destructor lifecycle implementation for the `A.FormBuilderII` class.
+     * Lifecycle.
      *
      * @method destructor
      * @protected
@@ -52,19 +50,19 @@ A.FormBuilderII  = A.Base.create('form-builder-II', A.Widget, [
     },
 
     /**
-     * Sets the form builder modal visibility to `false`.
+     * Hides the fields modal.
      *
      * @method hideFieldsPanel
      */
     hideFieldsPanel: function() {
-        this._modal.set('visible', false);
+        this._modal.hide();
     },
 
     /**
-     * Adds a new Field Type to the Form Buider.
+     * Adds a new Field Type to the Form Builder.
      *
      * @method registerFieldType
-     * @param val {Array | Object | A.FormBuilderIIFieldType}
+     * @param {Array | Object | A.FormBuilderIIFieldType} val
      */
     registerFieldType: function(val) {
         var instance = this;
@@ -80,25 +78,25 @@ A.FormBuilderII  = A.Base.create('form-builder-II', A.Widget, [
     },
 
     /**
-     * Sets the form builder modal visibility to `true`.
+     * Shows the fields modal.
      *
      * @method showFieldsPanel
      */
     showFieldsPanel: function() {
-        this._modal.set('visible', true);
+        this._modal.show();
     },
 
     /**
-     * Removes a Field Type from fieldTypes of this Form Buider.
+     * Removes a Field Type from fieldTypes of this Form Builder.
      *
      * @method unregisterFieldType
-     * @param val {String | A.FormBuilderIIFieldType}
+     * @param {A.FormBuilderIIFieldType} field
      */
     unregisterFieldType: function(field) {
         var index,
             fieldTypes = this.get('fieldTypes');
 
-        if (field instanceof A.FormBuilderIIFieldType) {
+        if (A.instanceOf(field, A.FormBuilderIIFieldType)) {
             index = fieldTypes.indexOf(field);
             if (index !== -1) {
                 fieldTypes[index].destroy();
@@ -111,22 +109,16 @@ A.FormBuilderII  = A.Base.create('form-builder-II', A.Widget, [
      * Removes a Field Type from fieldTypes of this Form Buider.
      *
      * @method _registerFieldType
-     * @param fieldObj {Object | A.FormBuilderIIFieldType}
+     * @param {Object | A.FormBuilderIIFieldType} field
      */
-    _registerFieldType: function(fieldObj) {
-        var fieldNode;
-
-        if (fieldObj instanceof A.FormBuilderIIFieldType) {
-            fieldNode = fieldObj;
-        }
-        else {
-            fieldNode = new A.FormBuilderIIFieldType(fieldObj);
+    _registerFieldType: function(field) {
+        if (!A.instanceOf(field, A.FormBuilderIIFieldType)) {
+            field = new A.FormBuilderIIFieldType(field);
         }
 
-        this.get('boundingBox').one('.field-types-list').appendChild(fieldNode.get('fieldTypeNode'));
-        this.get('fieldTypes').push(fieldNode);
+        this.get('boundingBox').one('.field-types-list').appendChild(field.get('node'));
+        this.get('fieldTypes').push(field);
     }
-
 }, {
 
     /**
@@ -140,15 +132,16 @@ A.FormBuilderII  = A.Base.create('form-builder-II', A.Widget, [
     ATTRS: {
 
         /**
-         * The collection of field types.
+         * The collection of field types that can be selected as fields for
+         * this form builder.
          *
          * @attribute fields
          * @default []
          * @type Array
          */
         fieldTypes: {
+            validator: A.Lang.isArray,
             value: []
         }
     }
-
 });