+++ src/aui-form-builder-II/HISTORY.md
@@ -4,4 +4,5 @@
 
 ## @VERSION@
 
+* [AUI-1647](https://issues.liferay.com/browse/AUI-1647) New form builder should know about the field types it will handle
 * [AUI-1658](https://issues.liferay.com/browse/AUI-1658) Create new Field Type submodule of Form Builder